refactor: #9 repository DI 적용 및 minor refactoring
