+++ delivery-service/src/main/java/com/pickple/delivery/domain/repository/DeliveryRepository.java
@@ -1,0 +1,9 @@
+package com.pickple.delivery.domain.repository;
+
+import com.pickple.delivery.domain.model.Delivery;
+
+public interface DeliveryRepository {
+
+    <S extends Delivery> S save(S entity);
+
+}
